# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `guest`, `superuser`, `hr_manager`, `project_manager`, `team_leader`, `team_member`, `compliance_officer`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `hr_manager`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created."
          }
        },
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Get a user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `hr_manager`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `hr_manager`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Delete a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/activities": {
      "get": {
        "operationId": "UsersController_getActivities",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role-Based Access Control",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Get activities for a user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "OfficeSync API",
    "description": "Backend API for the OfficeSync HR and PM Dashboard",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Role-Based-Access": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Enter your role to unlock the endpoints (e.g., platform_admin, superuser, hr_manager, team_leader, team_member)"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "example": "Arjun Mehta"
          },
          "email": {
            "type": "string",
            "example": "arjun@acme.test"
          },
          "password_hash": {
            "type": "string",
            "example": "Password123!"
          },
          "managerUserId": {
            "type": "string",
            "example": "user-uuid-123"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "role": {
            "type": "string",
            "example": "Team Member"
          },
          "phone": {
            "type": "string",
            "example": "+91 9876543210"
          },
          "companyId": {
            "type": "string",
            "example": "comp-uuid-123"
          },
          "branchId": {
            "type": "string",
            "example": "branch-uuid-456"
          }
        },
        "required": [
          "full_name",
          "email",
          "companyId"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "example": "Arjun Mehta"
          },
          "email": {
            "type": "string",
            "example": "arjun@acme.test"
          },
          "password_hash": {
            "type": "string",
            "example": "Password123!"
          },
          "managerUserId": {
            "type": "string",
            "example": "user-uuid-123"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "role": {
            "type": "string",
            "example": "Team Member"
          },
          "phone": {
            "type": "string",
            "example": "+91 9876543210"
          },
          "companyId": {
            "type": "string",
            "example": "comp-uuid-123"
          },
          "branchId": {
            "type": "string",
            "example": "branch-uuid-456"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "Role-Based-Access": []
    }
  ]
}